@@ -35,13 +35,16 @@
   },
   "devDependencies": {
     "babel-cli": "6.7.7",
+    "babel-eslint": "6.0.4",
     "babel-plugin-add-module-exports": "0.1.2",
     "babel-preset-es2015": "6.6.0",
     "babel-preset-react": "6.5.0",
     "babel-preset-react-hmre": "1.1.1",
     "babel-preset-stage-0": "6.5.0",
-    "eslint": "2.8.0",
-    "eslint-plugin-react": "5.0.1",
+    "eslint": "2.9.0",
+    "eslint-config-airbnb": "7.0.0",
+    "eslint-plugin-react": "4.3.0",
+    "eslint-plugin-jsx-a11y": "0.6.2",
     "feed": "0.3.0",
     "front-matter": "2.0.7",
     "markdown-it": "6.0.1",
