@@ -1,33 +1,38 @@
 module.exports = {
-  "env": {
-    "browser": true,
-    "commonjs": true,
-    "es6": true,
-    "node": true
+  env: {
+    browser: true,
+    commonjs: true,
+    es6: true,
+    node: true,
+    mocha: true,
   },
 
-  "parserOptions": {
-    "ecmaFeatures": {
-      "experimentalObjectRestSpread": true,
-      "jsx": true
-    },
-    "sourceType": "module"
-  },
-
-  "plugins": [
-    "react"
+  parser: 'babel-eslint',
+  plugins: [
+    'react'
   ],
-  "extends": "eslint:recommended",
+  extends: [
+    'airbnb'
+  ],
 
-  "rules": {
-    "indent": [2, 2],
-    "linebreak-style": [2, "unix"],
-    "quotes": [2, "single"],
-    "semi": [2, "always"],
-    "no-console": [0],
+  rules: {
+    strict: ['off'],
+    'no-console': ['off'],
+    'default-case': ['off'],
+    'no-param-reassign': ['off'],
+    'no-use-before-define': ['off'],
+    'consistent-return': ['off'],
 
-    "react/jsx-uses-vars": [2],
-    "react/jsx-uses-react": [2],
-    "react/display-name": [2]
+    'react/prefer-stateless-function': ['off'], // might reconsider this
+    'react/wrap-multilines': ['off'],
+
+    indent: ['error', 2, { SwitchCase: 1}],
+    'linebreak-style': ['error', 'unix'],
+    semi: ['error', 'always'],
+    'brace-style': ['error', 'stroustrup'],
+    'space-before-function-paren': ['error', 'never'],
+    'prefer-arrow-callback': ['error', { allowNamedFunctions: true }],
+    'comma-dangle': ['error', 'always-multiline'],
+    'func-names': ['error']
   }
 };
