@@ -7,6 +7,10 @@ import { rhythm } from 'utils/typography';
 import shortDate from 'utils/shortDate';
 
 export default class PostLink extends React.Component {
+  static propTypes = {
+    post: React.PropTypes.object.isRequired,
+  }
+
   render() {
     const post = this.props.post;
     const data = post.data;
@@ -14,14 +18,15 @@ export default class PostLink extends React.Component {
     return (
       <div
         style={{
-          marginBottom: rhythm(1/4)
+          marginBottom: rhythm(0.25),
         }}
       >
         <Link to={prefixLink(post.path)}>
           {data.title || post.path}
         </Link>
         <span className="date">
-          {' ' + shortDate(data.date)}
+          {' '}
+          {shortDate(data.date)}
         </span>
       </div>
     );
