@@ -10,7 +10,7 @@ import last from 'lodash/last';
 import { config } from 'config';
 
 
-function buildPiwikSetup({domainCDN, domainPiwik}) {
+function buildPiwikSetup({ domainCDN, domainPiwik }) {
   return `
     window._paq = window._paq || [];
     window._paq.push(['setTrackerUrl', '${domainPiwik}/piwik.php']);
@@ -24,8 +24,9 @@ function buildPiwikSetup({domainCDN, domainPiwik}) {
     window.start = new Date();
 
     // add tag
-    var d=document, g=d.createElement('script'), s=d.getElementsByTagName('script')[0]; g.type='text/javascript';
-    g.defer=true; g.async=true; g.src='${domainCDN}/js/piwik/2.15.0/piwik.min.js';
+    var d=document, g=d.createElement('script'), s=d.getElementsByTagName('script')[0];
+    g.type='text/javascript'; g.defer=true; g.async=true;
+    g.src='${domainCDN}/js/piwik/2.15.0/piwik.min.js';
     s.parentNode.insertBefore(g,s);
   `;
 }
@@ -34,6 +35,13 @@ const now = new Date();
 const buster = now.getTime();
 
 export default class HTML extends React.Component {
+  static propTypes = {
+    body: React.PropTypes.object,
+    title: React.PropTypes.object,
+    location: React.PropTypes.object,
+    routes: React.PropTypes.object,
+  }
+
   render() {
     const buildMode = Boolean(this.props.body);
 
@@ -45,10 +53,10 @@ export default class HTML extends React.Component {
 
     const metaTags = generateMetaTags(post, config, path);
 
-    const piwikSetup = buildPiwikSetup({domainCDN, domainPiwik});
+    const piwik = buildPiwikSetup({ domainCDN, domainPiwik });
     const bundle = <script async defer src={prefixLink(`/bundle.js?t=${buster}`)} />;
     const stylesheet =
-      <link rel='stylesheet' href={prefixLink(`/styles.css?t=${buster}`)} />;
+      <link rel="stylesheet" href={prefixLink(`/styles.css?t=${buster}`)} />;
 
     let js = bundle;
     let css = null;
@@ -63,22 +71,23 @@ export default class HTML extends React.Component {
       <html lang="en">
         <head>
           <title>{title}</title>
-          <meta charSet="utf-8"/>
-          <meta httpEquiv="X-UA-Compatible" content="IE=edge"/>
-          <meta name='viewport' content='initial-scale=1.0'/>
+          <meta charSet="utf-8" />
+          <meta httpEquiv="X-UA-Compatible" content="IE=edge" />
+          <meta name="viewport" content="initial-scale=1.0" />
           {metaTags}
-          <link rel="shortcut icon" href={favicon}/>
-          <TypographyStyle/>
+          <link rel="shortcut icon" href={favicon} />
+          <TypographyStyle />
           {css}
         </head>
         <body className="landing-page">
-          <div id="react-mount" dangerouslySetInnerHTML={{__html: this.props.body}} />
+          <div id="react-mount" dangerouslySetInnerHTML={{ __html: this.props.body }} />
           {js}
-          <script type="text/javascript" dangerouslySetInnerHTML={{__html: piwikSetup}} />
+          <script type="text/javascript" dangerouslySetInnerHTML={{ __html: piwik }} />
           <noscript>
             <img
               src={`${domainPiwik}/piwik.php?idsite=3&rec=1&url=${encodedPath}`}
-              style={{border: 0}}
+              style={{ border: 0 }}
+              alt="tracker"
             />
           </noscript>
         </body>
