@@ -14,6 +14,10 @@ import { config } from 'config';
 
 
 export default class Popular extends React.Component {
+  static propTypes = {
+    route: React.PropTypes.object.isRequired,
+  }
+
   render() {
     const title = 'Popular Posts';
 
@@ -25,7 +29,7 @@ export default class Popular extends React.Component {
 
     const pageLinks = map(posts, post => (
       <li key={post.path}>
-        <TextPreview  post={post} />
+        <TextPreview post={post} />
       </li>
     ));
 
@@ -39,7 +43,7 @@ export default class Popular extends React.Component {
           <hr
             style={{
               marginTop: rhythm(2),
-              marginBottom: rhythm(1)
+              marginBottom: rhythm(1),
             }}
           />
           <Author {...this.props} />
