@@ -9,7 +9,7 @@ import writeIfDifferent from '../utils/writeIfDifferent';
 
 
 const posts = loadPosts({
-  markdown: false
+  markdown: false,
 });
 const counts = getTagCounts(posts);
 
@@ -19,8 +19,8 @@ const templatePath = path.join(__dirname, '../components/_tagTemplate.js');
 const template = fs.readFileSync(templatePath).toString();
 const findTag = /"TAG"/m;
 
-_.forEach(_.keys(counts), function(tag) {
-  const filePath = path.join(__dirname, '../pages/tags', tag + '.js');
+_.forEach(_.keys(counts), tag => {
+  const filePath = path.join(__dirname, '../pages/tags', `${tag}.js`);
   const contents = template.replace(findTag, `"${tag}"`);
 
   writeIfDifferent(filePath, contents);
