@@ -4,10 +4,10 @@ import loadPosts from '../utils/loadPosts';
 import writeIfDifferent from '../utils/writeIfDifferent';
 
 
-const limit = parseInt(process.argv[2]) || 1;
+const limit = parseInt(process.argv[2], 10) || 1;
 const posts = loadPosts({
   limit,
-  markdown: false
+  markdown: false,
 });
 
 function removeSmartQuotes(value) {
@@ -16,7 +16,7 @@ function removeSmartQuotes(value) {
     .replace(/[“”]/g, '"');
 }
 
-_.forEach(posts, function(post) {
+_.forEach(posts, post => {
   console.log('checking', post.path);
 
   const contents = removeSmartQuotes(post.contents);
