@@ -4,13 +4,13 @@ import loadPosts from '../utils/loadPosts';
 import writeIfDifferent from '../utils/writeIfDifferent';
 
 
-const limit = parseInt(process.argv[2]) || 1;
+const limit = parseInt(process.argv[2], 10) || 1;
 const posts = loadPosts({
   limit,
-  markdown: false
+  markdown: false,
 });
 
-_.forEach(posts, function(post) {
+_.forEach(posts, post => {
   console.log('checking', post.path);
 
   const dupeLink = /\[([^\)]+)\]\(\1\)/g;
