@@ -4,11 +4,12 @@ import sortBy from 'lodash/sortBy';
 import fromPairs from 'lodash/fromPairs';
 import toPairs from 'lodash/toPairs';
 
+
 export default function getTagCounts(posts) {
   const lookup = Object.create(null);
 
-  forEach(posts, function(post) {
-    forEach(post.data.tags, function(tag) {
+  forEach(posts, post => {
+    forEach(post.data.tags, tag => {
       const list = lookup[tag] = lookup[tag] || [];
       list.push(post);
     });
@@ -16,14 +17,8 @@ export default function getTagCounts(posts) {
 
   let counts = toPairs(lookup);
 
-  counts = map(counts, function(array) {
-    return [array[0], array[1].length];
-  });
-
-  counts = sortBy(counts, function(array) {
-    return array[1];
-  });
-
+  counts = map(counts, array => [array[0], array[1].length]);
+  counts = sortBy(counts, array => array[1]);
   counts.reverse();
 
   return fromPairs(counts);
