@@ -1,16 +1,15 @@
 import moment from 'moment';
 
 export default function shortDate(date) {
-  var now = moment(new Date());
-  var instance = moment(date);
+  const now = moment(new Date());
+  const instance = moment(date);
 
-  var currentYear = now.format('YYYY');
-  var year = instance.format('YYYY');
+  const currentYear = now.format('YYYY');
+  const year = instance.format('YYYY');
 
   if (currentYear === year) {
     return instance.format('MMM DD');
   }
-  else {
-    return instance.format('YYYY MMM DD');
-  }
+
+  return instance.format('YYYY MMM DD');
 }
